requirements.txt
# Core Dependencies
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.5

# Data Processing (Optional but recommended)
pandas==2.0.1
numpy==1.24.3

# Web Framework Options (Choose one if building web interface)
flask==2.3.2          # Traditional web framework
fastapi==0.95.2       # Modern async web framework
streamlit==1.22.0     # Quick web app creation

# Development and Testing
pytest==7.3.1         # Unit testing
flake8==6.0.0         # Code linting
black==23.3.0         # Code formatting

# API Interaction
tmdbsimple==2.9.1     # Optional TMDB wrapper
#watchmode==0.1.0      # Optional WatchMode API helper